+++ colabdesign/af/alphafold/data/mmcif_parsing.py
