+++ colabdesign/af/alphafold/data/tools/hhblits.py
@@ -20,7 +20,7 @@ import subprocess
 from typing import Any, Mapping, Optional, Sequence
 
 from absl import logging
-from colabdesign.alphafold.data.tools import utils
+from colabdesign.af.alphafold.data.tools import utils
 # Internal import (7716).
