+++ colabdesign/af/alphafold/data/tools/hhsearch.py
@@ -21,7 +21,7 @@ from typing import Sequence
 
 from absl import logging
 
-from colabdesign.alphafold.data.tools import utils
+from colabdesign.af.alphafold.data.tools import utils
 # Internal import (7716).
