+++ colabdesign/af/alphafold/data/tools/hmmbuild.py
@@ -19,7 +19,7 @@ import re
 import subprocess
 
 from absl import logging
-from colabdesign.alphafold.data.tools import utils
+from colabdesign.af.alphafold.data.tools import utils
 # Internal import (7716).
