+++ colabdesign/af/alphafold/data/tools/jackhmmer.py
@@ -23,7 +23,7 @@ from urllib import request
 
 from absl import logging
 
-from colabdesign.alphafold.data.tools import utils
+from colabdesign.af.alphafold.data.tools import utils
 # Internal import (7716).
