+++ colabdesign/af/alphafold/model/tf/shape_placeholders.py
