+++ colabdesign/af/loss.py
@@ -6,7 +6,7 @@ from colabdesign.af.misc import _np_get_6D_loss, _np_rmsd, _np_kabsch
 from colabdesign.af.misc import update_seq, update_aatype 
 from colabdesign.af.misc import get_plddt, get_pae, get_fape_loss, get_dgram_loss, get_rmsd_loss_w, get_sc_rmsd
 
-from colabdesign.alphafold.model import model, folding, all_atom
+from colabdesign.af.alphafold.model import model, folding, all_atom
 
 ####################################################
 # AF_LOSS - setup loss function
