+++ colabdesign/af/utils.py
@@ -3,7 +3,7 @@ import jax.numpy as jnp
 import numpy as np
 
 from colabdesign.af.misc import jalview_color_list, _np_kabsch, order_restype
-from colabdesign.alphafold.common import protein
+from colabdesign.af.alphafold.common import protein
 
 # import matplotlib
 import matplotlib
