+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import setup, find_packages
 setup(
     name='ColabDesign',
     version='1.0.1',
-    packages=find_packages(include=['colabdesign/*']),
+    packages=find_packages(include=['colabdesign*']),
     install_requires=['py3Dmol','absl-py','biopython',
                       'chex','dm-haiku','dm-tree',
                       'immutabledict','jax','ml-collections',
